requirements.txt
beautifulsoup4==4.8.1
Django==2.2.7
django-rest-framework==0.1.0
djangorestframework==3.10.3
psycopg2==2.8.4
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
